+++ backend/canisters/local_user_index/impl/src/updates/join_channel.rs
@@ -7,8 +7,11 @@ use local_user_index_canister::join_channel::{Response::*, *};
 #[update(guard = "caller_is_openchat_user", msgpack = true)]
 #[trace]
 async fn join_channel(args: Args) -> Response {
-    let user_details =
-        mutate_state(|state| state.get_calling_user_and_process_credentials(args.verified_credential_args.as_ref()));
+    let (user_details, previous_user_ids) = mutate_state(|state| {
+        let user_details = state.get_calling_user_and_process_credentials(args.verified_credential_args.as_ref());
+        let previous_user_ids = state.data.migrated_user_ids.previous_ids(user_details.user_id);
+        (user_details, previous_user_ids)
+    });
 
     let is_bot = user_details.user_type.is_bot();
 
@@ -25,6 +28,7 @@ async fn join_channel(args: Args) -> Response {
         unique_person_proof: user_details.unique_person_proof.clone(),
         total_chit_earned: user_details.chit.total_earned,
         composite_gate_index: args.composite_gate_index,
+        previous_user_ids,
     };
     match community_canister_c2c_client::c2c_join_channel(args.community_id.into(), &c2c_args).await {
         Ok(response) => match response {